src/main/java/com/baldurtech/idCard/IdCardRepository.java
package com.baldurtech.idCard;

import javax.persistence.*;
import org.springframework.stereotype.Repository;
import org.springframework.transaction.annotation.Transactional;

@Repository
@Transactional
public class IdCardRepository {
    @PersistenceContext
    EntityManager entityManager;
    
    public void setEntityManager(EntityManager entityManager) {
        this.entityManager = entityManager;
    }
    
    public IdCard save(IdCard idCard) {
        entityManager.persist(idCard);
        return idCard;
    }
}